requirements.txt
fastapi
langchain
langchain-community
openai
langdetect
deep-translator
python-dotenv
wikipedia
arxiv
duckduckgo-search
ddgs
requests
google-cloud-texttospeech
flask
uvicorn[standard]
gunicorn
starlette
langchain_openai
gtts
pydantic
pydub
